refactor(animations): move `BrowserAnimationBuilder` to the animation package.
`AnimationBuilder` has no providers with `provideAnimationsAsync` because it was located in `@angular/plaform-browser/animations`. Since any import from `@angular/plaform-browser/animations` breaks the lazy-loading of the animation code, we have to move this class outside of it => to the animation package.

`@angular/plaform-browser/animations` will now import `BrowserAnimationBuilder` from `@angular/animations/browser` to maintain the compatibility.

Now, if you want to use an `AnimationBuilder` with `provideAnimationsAsync()`, please inject directly `BrowserAnimationBuilder`, it is now public and `providedIn:'root'`.

fixes #52096